Add team suggestion functionality

+++ app/src/main/java/junction/brunz/data/model/user/UserModel.kt
@@ -25,7 +25,7 @@ data class UserModel(
   @SerializedName("religion") val religion: String = "none",
   @SerializedName("activity") val activity: String? = "none",
   @SerializedName("color") val color: String = "green",
-  @SerializedName("weight") val weight: Double = 65.0,
+  @SerializedName("weight") val weight: Int = 65,
   @SerializedName("height") val height: Double = 1.8,
   @SerializedName("budget") val budget: String? = "medium",
   @SerializedName("cuisine") val cuisine: String = "",
